// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  SUPER_ADMIN: 'SUPER_ADMIN',
  SCHOOL_ADMIN: 'SCHOOL_ADMIN',
  TEACHER: 'TEACHER',
  STUDENT: 'STUDENT',
  PARENT: 'PARENT',
  INDIVIDUAL_LEARNER: 'INDIVIDUAL_LEARNER'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const AssessmentType = {
  TOPIC_BASED: 'TOPIC_BASED',
  WEEKLY: 'WEEKLY',
  TERMLY: 'TERMLY',
  YEARLY: 'YEARLY'
} as const

export type AssessmentType = (typeof AssessmentType)[keyof typeof AssessmentType]